refactor: fix schema
