+++ src/services/item/index.ts
@@ -19,7 +19,7 @@ import graaspApps from './plugins/app';
 import graaspEmbeddedLinkItem from './plugins/embeddedLink';
 import { PREFIX_EMBEDDED_LINK } from './plugins/embeddedLink/service';
 import graaspEnrollPlugin from './plugins/enroll';
-import graaspEtherpadPlugin from './plugins/etherpad';
+import graaspEtherpadPlugin from './plugins/etherpad/controller';
 import graaspFileItem from './plugins/file';
 import itemGeolocationPlugin from './plugins/geolocation/index';
 import graaspH5PPlugin from './plugins/html/h5p';
