--- src/services/item/plugins/etherpad/index.ts
@@ -1,2 +1,0 @@
-export { default } from './service-api';
-export * from './types';
