+++ src/services/item/plugins/etherpad/schemas.ts
@@ -1,4 +1,11 @@
+import { Type } from '@sinclair/typebox';
+import { StatusCodes } from 'http-status-codes';
+
+import { FastifySchema } from 'fastify';
+
 import { customType } from '../../../../plugins/typebox';
+import { errorSchemaRef } from '../../../../schemas/global';
+import { itemSchemaRef } from '../../schemas';
 
 export const createEtherpad = {
   operationId: 'createEtherpad',
@@ -7,12 +14,16 @@ export const createEtherpad = {
   description: 'Create an etherpad item.',
 
   querystring: customType.StrictObject({
-    parentId: customType.UUID(),
+    parentId: Type.Optional(customType.UUID()),
   }),
   body: customType.StrictObject({
     name: customType.ItemName(),
   }),
-};
+  response: {
+    [StatusCodes.OK]: itemSchemaRef,
+    '4xx': errorSchemaRef,
+  },
+} as const satisfies FastifySchema;
 
 export const getEtherpadFromItem = {
   operationId: 'getEtherpadFromItem',
@@ -21,9 +32,18 @@ export const getEtherpadFromItem = {
   description: 'Get etherpad information from item id',
 
   querystring: customType.StrictObject({
-    mode: customType.EnumString(['read', 'write']),
+    mode: Type.Optional(customType.EnumString(['read', 'write'])),
   }),
   params: customType.StrictObject({
     itemId: customType.UUID(),
   }),
+  response: {
+    [StatusCodes.OK]: customType.StrictObject(
+      { padUrl: Type.String() },
+      {
+        description: 'Successful Response',
+      },
+    ),
+    '4xx': errorSchemaRef,
+  },
 };
